Выполняет задание 4.1
